MasterFile_Class(code-CURD)

+++ src/main/java/com/pccw/backend/repository/BaseRepository.java
@@ -20,11 +20,11 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @NoRepositoryBean
-public interface BaseRepository<T> extends JpaRepository<T,Long>,JpaSpecificationExecutor<T> {
+    public interface BaseRepository<T> extends JpaRepository<T,Integer>,JpaSpecificationExecutor<T> {
 
     // Page<DbResRight> findAll(Specification<DbResRight> spec,Pageable p);
 
     @Modifying
     @Transactional
-    long deleteByIdIn(ArrayList<Long> ids);
+    long deleteByIdIn(ArrayList<Integer> ids);
 }
